Minor change to make framework build succesfully in msvc14, plus an improvement in the instructions for creation of a certificate for wws testing

+++ 3FD/web_wws_utils.cpp
@@ -24,7 +24,7 @@ namespace _3fd
 			{
 				return WS_XML_STRING
 				{
-					str.length(),
+					static_cast<ULONG> (str.length()),
 					(BYTE *)str.data(),
 					nullptr,
 					0
@@ -45,7 +45,7 @@ namespace _3fd
 				void *ptr = heap.Alloc<WS_XML_STRING>();
 				return new (ptr) WS_XML_STRING
 				{
-					str.length(),
+					static_cast<ULONG> (str.length()),
 					(BYTE *)str.data(),
 					nullptr,
 					0
